Pylint cleanup C0302
